Decouple view switch and view render in backlog
