+++ frontend/src/App.css
@@ -3,5 +3,5 @@
 }
 
 body {
-  background-color: #F7F7F7;
+  background-color: white;
 }
