+++ frontend/src/components/Backlog/BacklogDay.css
@@ -1,9 +1,7 @@
 .BacklogDay {
   margin: 0.5em;
-  background-color: white;
-  box-shadow: 0 3px 1px -2px rgba(0, 0, 0, .2),
-  0 2px 2px 0 rgba(0, 0, 0, .14),
-  0 1px 5px 0 rgba(0, 0, 0, .12);
+  background-color: #F9F9F9;
+  box-shadow: 0 1px 1px -1px rgba(0, 0, 0, .1), 0 2px 2px 0 rgba(0, 0, 0, .1), 0 1px 5px 0 rgba(0, 0, 0, .11);
   height: 250px;
   overflow: hidden; /* TODO fix this hack */
 }
